refactor(api): move auth code from deps.py to stitch.api.auth module
