+++ deployments/api/src/stitch/api/auth.py
