+++ deployments/api/src/stitch/api/routers/resources.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter
 
 from stitch.api.db import resource_actions
 from stitch.api.db.config import UnitOfWorkDep
-from stitch.api.deps import CurrentUser
+from stitch.api.auth import CurrentUser
 from stitch.api.entities import CreateResource, Resource
